requirements.txt
python-xlib==0.33
ewmh==0.1.6
